Rework how compilations are handled
Testing over a large library reveals that the COMPILATION flag is
usually misleading. Sometimes it's set incorrectly (as is the case for
anthologies), sometimes it's not set when it should be (the tagger
forgot it)

The new algorithm ignore the COMPILATION tag by default and only uses
the ALBUMARTIST and ARTIST tags. If the ARTISTs are all the same, the
album is not a compilation. If the ALBUMARTIST is set, it's not a
compilation. In other cases, it is considered a compilation.

The file template will now always include the ARTIST if any of the
artists differ, regardless of if it's a compilation or not.

Setting `is_compilation` across an album will now force the path
template for that album, regardless of the ALBUMARTIST or ARTIST tags.
If a library properly uses the COMPILATION tag, an override can be used
to force songs with COMPILATION=1 to set the `is_compilation` flag.